backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
openai==1.3.0
supabase==2.0.0
pydantic==2.5.0
httpx==0.24.1  # <- supabase 요구 조건에 맞는 버전으로 수정
python-multipart==0.0.6
googlemaps==4.10.0